Fix handling of parameters with spaces
If parameters are enclosed in " and contain spaces they can only
be passed on properly if the quoting is handled correctly. This can
be done for example by using "$@" in all places. Simply assigning
"$@" to a varible and using that variable will break the quoting.
For example "1" "2 2" will become "1 2 2" or 1 2 3 depending on what you
do.