detects busybox 1.30.0+ with updated timeout flags
Busybox updated timeout's flags to match coreutils.
This broke the script that was specifically dealing
with BusyBox being different from coreutils.

To fix the problem without breaking support for older
versions of busy box, this a version check of busybox
to determine which version of the flags to use.

Note, the changes use bash's regular expression's BASH_REMATCH
to get the job done.  For Alpine users, this means you'll have
to install bash and execute wait-for-it using bash.