fix: Change timeout invocation to support breaking change
In busybox commit [c9720a761][], a breaking change was introduced that
removes the `-t` flag to `timeout` in favor of just taking the time as
the second arg, bringing it in line with coreutils' `timeout`. This
commit was included in tag [1.31.1][], released on 25 Oct, 2019, at
which point my integration tests that spin up a server using wait-for-it
started to fail.

[c9720a761]: https://github.com/mirror/busybox/commit/c9720a761e88e83265b4d75808533cdfbc66075b
[1.31.1]: https://github.com/mirror/busybox/releases/tag/1_31_1